revert original exercise 2
