+++ src/exercise/02.js
@@ -3,35 +3,19 @@
 
 import * as React from 'react'
 
-function useLocalStorageState(
-  key,
-  defaultValue = '',
-  {serialize = JSON.stringify, deserialize = JSON.parse} = {},
-) {
-  const [state, setState] = React.useState(() => {
-    const valueInLocalStorage = window.localStorage.getItem(key) ?? defaultValue
-    if (valueInLocalStorage) {
-      return deserialize(valueInLocalStorage)
-    }
-    return typeof defaultValue === 'function' ? defaultValue() : defaultValue
-  })
-
-  const prevKeyRef = React.useRef(key)
-
-  React.useEffect(() => {
-    const prevKey = prevKeyRef.current
-    if (prevKey !== key) {
-      window.localStorage.removeItem(prevKey)
-    }
-    prevKeyRef.current = key
-    window.localStorage.setItem(key, serialize(state))
-  }, [key, serialize, state])
-
-  return [state, setState]
-}
-
 function Greeting({initialName = ''}) {
-  const [name, setName] = useLocalStorageState('name', initialName)
+  // 🐨 initialize the state to the value from localStorage
+  // 💰 window.localStorage.getItem('name') ?? initialName
+  const [name, setName] = React.useState(
+    window.localStorage.getItem('name') ?? initialName,
+  )
+
+  // 🐨 Here's where you'll use `React.useEffect`.
+  // The callback should set the `name` in localStorage.
+  // 💰 window.localStorage.setItem('name', name)
+  React.useEffect(() => {
+    window.localStorage.setItem('name', name)
+  })
 
   function handleChange(event) {
     setName(event.target.value)
